proyecto

+++ proyecto/menu.py
@@ -14,24 +14,53 @@ def menu():
             if curso_op== 'a':
                 curso.getCurso()
             elif curso_op== 'b':
-                pass
+                id_consulta= input("Ingrese el ID que desea obtener la información:\n")
+                if  id_consulta.isnumeric():
+                    curso.getCurso_ID(id_consulta)
+                else:
+                    print("Debe ingresar un ID numérico")
             elif curso_op== 'c':
                 actualizar_op= input("Indique lo que desea actualizar:\n\ta)Nombre\n\tb)Tiempo\n\tc)Todos los datos")
                 actualizar_op.lower()
                 if actualizar_op== 'a':
-                    pass
+                    id_actualizaName= input("Ingrese el ID que desea actualizar:\n")
+                    if id_actualizaName.isnumeric():
+                        nombre_Name= input("Escriba el nombre:\n")
+                        curso.updateCurso_ID(id_actualizaName, nombre_Name)
+                    else:
+                        print("Debe ingresar un ID numérico")
                 elif actualizar_op== 'b':
-                    pass
+                    id_actualizaTime= input("Ingrese el ID que desea actualizar:\n")
+                    if id_actualizaTime.isnumeric():
+                        tiempo_Time= input("Indique el nuevo tiempo:\n")
+                        curso.updateCurso_TiempoID(id_actualizaTime, tiempo_Time)
+                    else:
+                        print("Debe ingresar un ID numérico")
                 elif actualizar_op== 'c':
-                    pass
+                    id_actualizaTotal= input("Ingrese el ID que desea actualizar:\n")
+                    if id_actualizaTotal.isnumeric():
+                        nombre= input("Escriba el nombre:\n")
+                        descripcion= input("Escriba la nueva descripcion:\n")
+                        tiempo= input("Indique el nuevo tiempo:\n")
+                        usuario= input("Escriba el usuario:\n")
+                        curso.updateCurso_TotalID(id_actualizaTotal, nombre, descripcion, tiempo, usuario)
+                    else:
+                        print("Debe ingresar un ID numérico")
                 else:
                     print("Error, ingrese una opcion válida")
                 pass
             elif curso_op== 'd':
-                pass
+                new_name= input("Ingrese el nombre del usuario que quiere crear:\n")
+                new_desc= input("Ingrese la descripcion:\n")
+                new_time= input("Ingrese el tiempo:\n")
+                new_usuario= input("Ingrese el usuario:\n")
+                curso.createCurso(new_name, new_desc, new_time, new_usuario)
             elif curso_op== 'e':
                 delete_id= input("Ingrese el ID que desea eliminar: ")
-                curso.deleteCurso(delete_id)
+                if delete_id.isnumeric():
+                    curso.deleteCurso(delete_id)
+                else:
+                    print("Debe ingresar un ID numérico")
             else:
                 print("Debe ingresar una opción correcta")
         elif opcion==2:
@@ -63,6 +92,7 @@ def menu():
                 print("Debe ingresar una opción correcta")
         elif opcion==3:
             print("Profesor")
+            profe=P.Database()
             profe_op= input("Indique la opción de lo que desea hacer:\n\ta)Consulta datos general\n\tb)Consulta datos individual por ID\n\tc)Actualiza un registro por medio del ID\n\td)Crea un registro\n\te)Elimina un registro por ID\n")
             profe_op.lower()
             if profe_op== 'a':
@@ -89,28 +119,33 @@ def menu():
                 print("Debe ingresar una opción correcta")
         elif opcion==4:
             print("Grupo")
+            group=G.Database()
             grupo_op= input("Indique la opción de lo que desea hacer:\n\ta)Consulta datos general\n\tb)Consulta datos individual por ID\n\tc)Actualiza un registro por medio del ID\n\td)Crea un registro\n\te)Elimina un registro por ID\n")
             grupo_op.lower()
             if grupo_op== 'a':
-                pass
+                group.getGroup()
             elif grupo_op== 'b':
-                pass
-            elif grupo_op== 'c':
-                actualizar_op= input("Indique lo que desea actualizar:\n\ta)Nombre\n\tb)Tiempo\n\tc)Todos los datos")
-                actualizar_op.lower()
-                if actualizar_op== 'a':
-                    pass
-                elif actualizar_op== 'b':
-                    pass
-                elif actualizar_op== 'c':
-                    pass
+                id_consulta= input("Ingrese el ID que desea obtener la información:\n")
+                if  id_consulta.isnumeric():
+                    group.getGroup_ID(id_consulta)
                 else:
-                    print("Error, ingrese una opcion válida")
-                pass
+                    print("Debe ingresar un ID numérico")
+            elif grupo_op== 'c':
+                id_actualiza= input("Ingrese el ID que desea actualizar:\n")
+                if id_actualiza.isnumeric():
+                    nombreGroup= int("Ecriba el nombre:\n")
+                    group.uptadeGroup_ID(id_actualiza, nombreGroup)
+                else:
+                    print("Debe ingresar un ID numérico")
             elif grupo_op== 'd':
-                pass
+                new_groupname=("Ingrese el nuevo nombre:\n")
+                group.createGroup(new_groupname)
             elif grupo_op== 'e':
-                pass
+                delete_id= input("Ingrese el ID que desea eliminar: ")
+                if delete_id.isnumeric():
+                    group.deleteCurso(delete_id)
+                else:
+                    print("Debe ingresar un ID numérico")
             else:
                 print("Debe ingresar una opción correcta")
         elif opcion==5:
